move: contentful.ts 폴더 변경(utils -> api)
